[#23] [FIX] 퀴즈 생성 서비스단에서 트랜잭션으로 처리

+++ src/main/java/com/mashup/munggoo/quiz/service/QuizService.java
@@ -2,13 +2,16 @@ package com.mashup.munggoo.quiz.service;
 
 import com.mashup.munggoo.exception.ConflictException;
 import com.mashup.munggoo.exception.NotFoundException;
+import com.mashup.munggoo.highlight.Highlight;
 import com.mashup.munggoo.quiz.domain.Quiz;
 import com.mashup.munggoo.quiz.dto.AnswerDto;
 import com.mashup.munggoo.quiz.dto.ReqResultDto;
 import com.mashup.munggoo.quiz.dto.ScoreDto;
+import com.mashup.munggoo.quiz.quizGenerator.QuizGenerator;
 import com.mashup.munggoo.quiz.repository.QuizRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -17,18 +20,30 @@ import java.util.List;
 @Service
 public class QuizService {
     private final QuizRepository quizRepository;
+    private final HighlightForQuizService highlightForQuizService;
 
-    public void delete(Long fileId){
+    @Transactional
+    public List<Quiz> createQuiz(Long fileId){
+        List<Highlight> highlights = highlightForQuizService.getHighlights(fileId);
+        if(highlights.isEmpty()){
+            throw new NotFoundException("Highlight Does Not Exist.");
+        }
+
         if(quizAlreadyExist(fileId)){
             quizRepository.deleteQuizzesByFileId(fileId);
         }
-    }
-    private boolean quizAlreadyExist(Long fileId){
-        return quizRepository.existsQuizzesByFileId(fileId);
+
+        List<Quiz> quizzes= QuizGenerator.generateQuizSet(highlights);
+
+        if(quizzes.isEmpty()){
+            throw new NotFoundException("Quiz Does Not Generated.");
+        }
+
+        return quizRepository.saveAll(quizzes);
     }
 
-    public List<Quiz> save(List<Quiz> quizList){
-        return quizRepository.saveAll(quizList);
+    private boolean quizAlreadyExist(Long fileId){
+        return quizRepository.existsQuizzesByFileId(fileId);
     }
 
     public List<Quiz> getQuiz(Long fileId){
